requirements.txt
asgiref==3.8.1
black==25.1.0
click==8.1.8
colorama==0.4.6
Django==5.1.7
django-cacheops==7.1
djangorestframework==3.15.2
drf-extensions==0.7.1
exceptiongroup==1.2.2
funcy==2.0
iniconfig==2.1.0
mypy-extensions==1.0.0
packaging==24.2
pathspec==0.12.1
pillow==11.0.0
platformdirs==4.3.7
pluggy==1.5.0
pytest==8.3.5
pytest-django==4.10.0
redis==5.2.1
sqlparse==0.5.3
tomli==2.2.1
typing_extensions==4.12.2
tzdata==2025.1
